Make SkAlignTo templated; optimize SkAlignNonPow2
Updates several structs in Graphite to be consistent about using
uint32_t for size/offset/alignment since they are post-validation
types where we know that uint32_t is sufficient. This is needed so
that the templated SkAlignTo and SkAlignNonPow2 can detect T correctly.

Bug: b/263130762
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1064297
Commit-Queue: Michael Ludwig <michaelludwig@google.com>

+++ src/gpu/ganesh/d3d/GrD3DCaps.cpp
@@ -927,7 +927,7 @@ GrCaps::SupportedWrite GrD3DCaps::supportedWritePixelsColorType(
 
     // Any buffer data needs to be aligned to 512 bytes and that of a single texel.
     size_t offsetAlignment = SkAlignTo(GrDxgiFormatBytesPerBlock(dxgiFormat),
-                                       D3D12_TEXTURE_DATA_PLACEMENT_ALIGNMENT);
+                                       (size_t) D3D12_TEXTURE_DATA_PLACEMENT_ALIGNMENT);
 
     const auto& info = this->getFormatInfo(dxgiFormat);
     for (int i = 0; i < info.fColorTypeInfoCount; ++i) {
